Blog Schema Service some checks add and criteria add enums

+++ src/main/java/elbar/company/learn_center_rest/service/blog/blog/BlogServiceImpl.java
@@ -16,9 +16,11 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.webjars.NotFoundException;
 
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
 @Service
@@ -37,7 +39,11 @@ public class BlogServiceImpl extends AbstractService<BlogValidator, BlogMapper, 
     @Override
     public ResponseEntity<Data<Void>> update(BlogUpdateDTO DTO) {
         validator.validOnUpdate(DTO);
-        Blog blog = repository.getByCode(DTO.getCode());
+        Optional<Blog> optional = repository.getByCode(DTO.getCode());
+        if (optional.isEmpty()) {
+            throw new NotFoundException("Blog not found");
+        }
+        Blog blog = optional.get();
         blog.setTitle(DTO.getTitle());
         blog.setDescription(DTO.getDescription());
         blog.setCategoryId(DTO.getCategoryId());
@@ -52,6 +58,10 @@ public class BlogServiceImpl extends AbstractService<BlogValidator, BlogMapper, 
     @Override
     public ResponseEntity<Data<Void>> delete(UUID key) {
         validator.validateKey(key);
+        Optional<Blog> optional = repository.getByCode(key);
+        if (optional.isEmpty()) {
+            throw new NotFoundException("Blog not found");
+        }
         repository.deleteByCode(key);
         return new ResponseEntity<>(new Data<>(true), HttpStatus.OK);
     }
@@ -59,19 +69,18 @@ public class BlogServiceImpl extends AbstractService<BlogValidator, BlogMapper, 
     @Override
     public ResponseEntity<Data<BlogGetDTO>> get(UUID key) {
         validator.validateKey(key);
-        return new ResponseEntity<>(new Data<>(mapper.fromGetDTO(repository.getByCode(key))), HttpStatus.OK);
+        return new ResponseEntity<>(new Data<>(mapper.fromGetDTO(repository.getByCode(key).orElseThrow(() -> new NotFoundException("Blog not found")))), HttpStatus.OK);
     }
 
     @Override
     public ResponseEntity<Data<BlogDetailDTO>> detail(UUID key) {
         validator.validateKey(key);
-        return new ResponseEntity<>(new Data<>(mapper.fromDetailDTO(repository.getByCode(key))), HttpStatus.OK);
+        return new ResponseEntity<>(new Data<>(mapper.fromDetailDTO(repository.getByCode(key).orElseThrow(() -> new NotFoundException("Blog not found")))), HttpStatus.OK);
     }
 
     @Override
     public ResponseEntity<Data<List<BlogGetDTO>>> list(BlogCriteria criteria) {
-        PageRequest request = PageRequest.of(criteria.getPage(), criteria.getSize());
-        Page<Blog> all = repository.findAll(request);
-        return new ResponseEntity<>(new Data<>(mapper.fromGetListDTO(all.toList()), all.getSize()), HttpStatus.OK);
+        PageRequest request = PageRequest.of(criteria.getPage(), criteria.getSize(), criteria.getSort(), criteria.getFieldsEnum().getValue());
+        return new ResponseEntity<>(new Data<>(mapper.fromGetListDTO(repository.findAll(request).stream().toList()), repository.count()), HttpStatus.OK);
     }
 }
